requirements.txt
Django>=3.2,<5.0
psycopg2-binary>=2.9.1
stripe>=5.0
djangorestframework>=3.12
gunicorn>=20.0.4
python-decouple>=3.6
drf-yasg>=1.21.5